requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]